disable sanitizer checks and add prlimit_std test
